Issue 6663 - Fix NULL subsystem crash in JSON error logging
Description: Fixes crash in JSON error logging when subsystem is NULL.
Parametrized test case for better debugging.

Relates: https://github.com/389ds/389-ds-base/issues/6663

Reviewed by: @mreynolds389
